requirements.txt
numpy==1.21.6
pandas==1.3.5
pathspec==0.10.1
PyYAML==6.0
